requirements.txt
torch
torch-geometric
torchmetrics
scikit-learn
networkx
pyyaml
tqdm
sentence-transformers
scipy
numpy
matplotlib
openai
uvloop
aiohttp
